fixed application stop script to only kill if existent

+++ Server/deployment/application_stop.sh
@@ -1,3 +1,3 @@
 #!/bin/bash
 echo "Stopping any existing node servers"
-pkill node
+if pgrep myServer; then pkill node; fi
